lips/pyproject.toml
[project]
name = "cici-frontend"
version = "0.1.0"
description = "Frontend UI for cici voice and text personal assistant"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
